+++ CHANGELOG.md
@@ -1,5 +1,9 @@
-## Version 2 (2019-11-17)
+## Version 2
+#### 2019-11-17
 - Removed support for Apify Crawler (this product has been removed)
 - Webhooks work out of the box without need to change payload template (for tasks).
 - Input - `datasetOrExecutionId` changed to `datasetId`.
 - Added sections to input (only visual change)
+
+#### 2019-12-31
+- Fixed: Excess rows/columns were wrongly trimmed if the range was not the first sheet. May have caused removal of data in the first sheet.